change server to be testable concurrently and mockable (major #refactor) (bad commit practice)
The code is changed so that we can spin up multiple servers to run tests concurrently and be able to easily
close any server. Previously we used nodemon to spin up the server on a separate process which limited our
ability to mock or run concurrently and similar behavior.
Also in this architecture, each test only loads necessary modules (specifically only necessary router will
be installed on the server) so in theory, it should run faster.